requirements.txt
# PaperStream Requirements
# ========================

# Core Framework
fastmcp>=2.0.0
starlette>=0.27.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# HTTP/Async
httpx>=0.24.0
aiohttp>=3.8.0

# Authentication
PyJWT>=2.8.0  # JWT verification for Unity Auth

# BioBERT / Transformers (for embeddings)
transformers>=4.35.0
torch>=2.0.0
sentencepiece>=0.1.99

# Scientific/Numerical
numpy>=1.24.0
scipy>=1.11.0

# PDF Processing
PyMuPDF>=1.23.0  # fitz
Pillow>=10.0.0
pdf2image>=1.16.0

# Database
qdrant-client>=1.7.0

# Utilities
pyyaml>=6.0.0
tqdm>=4.66.0

# Optional: For faster inference
# accelerate>=0.24.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
